backend/requirements.txt
Flask
psycopg2-binary
pytest
requests
gunicorn
python-dotenv